change: reorganize code to match db schemas
