+++ pyproject.toml
@@ -5,6 +5,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "asyncpg>=0.29.0",
+    "email-validator>=2.2.0",
     "fastapi>=0.114.2",
     "greenlet>=3.1.0",
     "pydantic>=2.9.1",
