+++ src/anime_rest_api/db/__init__.py
@@ -3,27 +3,20 @@ from sqlalchemy.schema import CreateSchema
 
 from .connection import DatabaseConnection
 from .errors import InvalidDbConnectionStateError
-from .models import DB_METADATA
-from .models import Show
-from .models import ShowBase
-from .models import ShowCreate
-from .models import ShowRead
-from .models import ShowUpdate
+from .models.auth import AUTH_METADATA
+from .models.content import CONTENT_METADATA
 
 __all__ = [
     "DatabaseConnection",
-    "Show",
-    "ShowBase",
-    "ShowCreate",
-    "ShowRead",
-    "ShowUpdate",
-    "DB_METADATA",
     "InvalidDbConnectionStateError",
     "setup_db",
+    "CONTENT_METADATA",
+    "AUTH_METADATA",
 ]
 
 
 async def setup_db(conn: AsyncConnection) -> None:
     """Setup the database."""
-    await conn.execute(CreateSchema(DB_METADATA.schema, if_not_exists=True))
-    await conn.run_sync(DB_METADATA.create_all)
+    for meta in [CONTENT_METADATA, AUTH_METADATA]:
+        await conn.execute(CreateSchema(meta.schema, if_not_exists=True))
+        await conn.run_sync(meta.create_all)
