--- src/anime_rest_api/db/models/__init__.py
@@ -1,8 +1,0 @@
-from .base import DB_METADATA
-from .shows import Show
-from .shows import ShowBase
-from .shows import ShowCreate
-from .shows import ShowRead
-from .shows import ShowUpdate
-
-__all__ = ["DB_METADATA", "ShowBase", "Show", "ShowCreate", "ShowRead", "ShowUpdate"]
