+++ src/anime_rest_api/db/models/auth/__init__.py
@@ -1,0 +1,3 @@
+from .base import AUTH_METADATA
+
+__all__ = ["AUTH_METADATA"]
