+++ src/anime_rest_api/db/models/auth/base.py
@@ -1,0 +1,5 @@
+from sqlmodel import MetaData
+
+AUTH_METADATA = MetaData(schema="auth")
+
+__all__ = ["AUTH_METADATA"]
