--- src/anime_rest_api/db/models/base.py
@@ -1,5 +1,0 @@
-from sqlmodel import MetaData
-
-DB_METADATA = MetaData(schema="anime_data")
-
-__all__ = ["DB_METADATA"]
