+++ src/anime_rest_api/db/models/content/__init__.py
@@ -1,0 +1,15 @@
+from .base import CONTENT_METADATA
+from .shows import Show
+from .shows import ShowBase
+from .shows import ShowCreate
+from .shows import ShowRead
+from .shows import ShowUpdate
+
+__all__ = [
+    "CONTENT_METADATA",
+    "ShowBase",
+    "Show",
+    "ShowCreate",
+    "ShowRead",
+    "ShowUpdate",
+]
