+++ src/anime_rest_api/db/models/content/base.py
@@ -1,0 +1,5 @@
+from sqlmodel import MetaData
+
+CONTENT_METADATA = MetaData(schema="content")
+
+__all__ = ["CONTENT_METADATA"]
